fix(bubble): change elevation rendering logic

+++ core/src/main/java/net/flectone/pulse/module/message/bubble/model/Bubble.java
@@ -25,7 +25,7 @@ public class Bubble {
 
     private final long duration;
 
-    private final int elevation;
+    private final float elevation;
 
     private final float interactionHeight;
